pyproject.toml
[tool.poetry]
name = "christmas_jumper_classification"
version = "1"
description = ""
authors = ["gv9 <georgie.vaughan@ricardo.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
pandas = "^1.5.1"
tensorflow = "^2.10.0"
matplotlib = "^3.6.2"
Pillow = "^9.3.0"
jupyter = "^1.0.0"
python-splitter = "^0.0.3"
jupytext = "^1.14.1"
jmd-imagescraper = "^1.0.2"
opencv-python = "^4.6.0"
sklearn = "^0.0.post1"
scikit-learn = "^1.2.0"
torchvision = "^0.14.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
